Merge Commentary to Develop

+++ App/Controller/AdminController.php
@@ -18,7 +18,7 @@
         private $form;
         private $security;
         private $date;
-
+        private $isAdmin;
 
         public function __construct(){
             $this->adminModel = new AdminModel;
@@ -28,66 +28,70 @@
             $this->date = Date::GetDate();
 
             $this->security = new Security;
+
+            (bool) $this->isAdmin = (isset($_SESSION['isAdmin'])) ? $_SESSION['isAdmin'] : false;
         } 
 
-        public function index(){    
+        public function index(){
             $allPost = $this->adminModel->index();
 
             parent::Render('App/View/AdminView.php',array('post'=>$allPost));
         }
 
-        public function update($id=false){
 
-            if($id != false){
-                $getPost = $this->adminModel->updatePost($id);
-                $getPost = $getPost->fetch(); 
+         public function update($id=false){
+            if($this->isAdmin){
+                if($id != false){
+                    $getPost = $this->adminModel->updatePost($id);
+                    $getPost = $getPost->fetch(); 
 
-                $this->session = new Session(array('idPost'),array($id));
-                $this->session->getSession();
+                    $this->session = new Session(array('idPost'),array($id));
+                    $this->session->getSession();
 
-                parent::Render('App/View/AdminUpdateView.php',array("post"=>$getPost));
+                    parent::Render('App/View/AdminUpdateView.php',array("post"=>$getPost));
+                }
+
+                if(isset($_POST['title']) AND isset($_POST['chapo']) AND isset($_FILES['image']) AND isset($_POST['content'])){ 
+                    $title = htmlspecialchars($_POST['title']);
+                    $chapo = htmlspecialchars($_POST['chapo']);
+                    $image = htmlspecialchars($_FILES['image']['name']);
+                    $content = htmlspecialchars($_POST['content']);
+                    
+
+                    $author = $_SESSION['user_id'];
+                    $idPost = $_SESSION['idPost'];
+                    
+                    $this->adminModel->updatePost($idPost,$title,$chapo,$image,$content,$author,$this->date);
+
+                    header('location: '.WebSiteLink.'admin/index');
+                }
             }
-
-            if(isset($_POST['title']) AND isset($_POST['chapo']) AND isset($_FILES['image']) AND isset($_POST['content'])){ 
- 
-                $title = htmlspecialchars($_POST['title']);
-                $chapo = htmlspecialchars($_POST['chapo']);
-                $image = htmlspecialchars($_FILES['image']['name']);
-                $content = htmlspecialchars($_POST['content']);
-
-                $author = $_SESSION['user_id'];
-                $idPost = $_SESSION['idPost'];
-                var_dump($title);
-                
-                $this->adminModel->updatePost($idPost,$title,$chapo,$image,$content,$author,$this->date);
-                //var_dump($this->adminModel->updatePost($title,$chapo,$image,$content,$author,$this->date,$idPost));
-
-                header('location: '.WebSiteLink.'admin/index');
-            }
-
-            //var_dump($this->adminModel->updatePost($title,$chapo,$image,$content,$author,$this->date,$id));             
         }
 
+
         public function add(){
-            if(isset($_POST['title']) AND isset($_POST['chapo']) AND isset($_FILES['image']) AND isset($_POST['content'])){ 
-                $title = htmlspecialchars($_POST['title']);
-                $chapo = htmlspecialchars($_POST['chapo']);
-                $image = htmlspecialchars($_FILES['image']['name']);
-                $content = htmlspecialchars($_POST['content']);
+            if($this->isAdmin){
+                if(isset($_POST['title']) AND isset($_POST['chapo']) AND isset($_FILES['image']) AND isset($_POST['content'])){ 
+                    $title = htmlspecialchars($_POST['title']);
+                    $chapo = htmlspecialchars($_POST['chapo']);
+                    $image = htmlspecialchars($_FILES['image']['name']);
+                    $content = htmlspecialchars($_POST['content']);
 
-                (int)$author = (int)$_SESSION['user_id'];
+                    (int)$author = (int)$_SESSION['user_id'];
 
-                $this->adminModel->addPost($title,$chapo,$image,$content,$author,$this->date);
+                    $this->adminModel->addPost($title,$chapo,$image,$content,$author,$this->date);
 
-                header('location: '.WebSiteLink.'admin/index');
+                    header('location: '.WebSiteLink.'admin/index');
+                }
             }
         }
 
         public function delete($id){
-            $this->adminModel->deletePost($id);
+            if($this->isAdmin){
+                $this->adminModel->deletePost($id);
 
-            header('location: '.WebSiteLink.'admin/index');
-            //parent::Render('App/View/AdminView.php',array());
+                header('location: '.WebSiteLink.'admin/index');
+            }
         }
 
     }
